Random Forest (v1) and all lakes
First complete attempt at modelling Random Forests for r_cyanobac and d_cyanobac (the ratio and difference of the cyanobacteria biovolume to the total biovolume of the phyoplankton).

Also, now all lakes contain all the code

Note: the linear regression doesn't work anymore, dure to a change in names. This will be updated in the future if this piece of code is still of interest